% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_removal.R
\name{cor_high_correlation_features}
\alias{cor_high_correlation_features}
\title{Determines which columns are too highly correlated.}
\usage{
cor_high_correlation_features(data, exclude_columns = character(0),
  threshold = 0.8)
}
\arguments{
\item{data}{The data frame to use for determining which columns are highly correlated. Must be all numeric.}

\item{exclude_columns}{Columns that will not be used to determine too highly correlated features. Should include at least your response.}

\item{threshold}{If the absolute correlation between two variables is higher than this value, the pair will be registered as too highly correlated. Defaults to 0.8}
}
\value{
A dataframe with two columns, representing pairs of columns that are highly correlated
}
\description{
Determines which columns are too highly correlated.
}
