% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_removal.R
\name{cor_remove_high_correlation_features}
\alias{cor_remove_high_correlation_features}
\title{Removes highly correlated features whilst keeping as many as possible, using heuristics.}
\usage{
cor_remove_high_correlation_features(train, exclude_columns = character(0),
  threshold = 0.8)
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{exclude_columns}{Columns that should not be considered for removal.}

\item{threshold}{If the correlation is between two columns is larger than this, both will be considered for removal}
}
\value{
A list of train set and the columns that were preserved
}
\description{
Removes highly correlated features whilst keeping as many as possible, using heuristics.
}
