% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{create_predict_function}
\alias{create_predict_function}
\title{Ensures a predict function can be called for custom pipeline functions}
\usage{
create_predict_function(.predict_function, ...)
}
\arguments{
\item{.predict_function}{The function that will be used for applying learned transformations to a new dataset.}

\item{...}{Arguments to \code{.predict_function}.}
}
\value{
A function that takes a \code{data} argument (for new data) and uses the provided arguments in \code{...}.
}
\description{
Ensures a predict function can be called for custom pipeline functions
}
\details{
The predict function of custom pipeline segments can cause problems when being redeployed if the predict function is not properly supplied to the function.
In these cases, this helper function can assist. It will force evaluation of its function argument before using it in a new function, ensuring no
delayed evaluation occurs. This should prevent errors along the likes of "predict function not found" when deploying the pipeline in a clean environment.
}
