% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{create_stats}
\alias{create_stats}
\title{Calculates stats based on custom functions on the response variable for each group provided in stat_cols.}
\usage{
create_stats(train, statistics_col, response, functions,
  too_few_observations_cutoff = 30)
}
\arguments{
\item{train}{The train dataset, as a data.table}

\item{statistics_col}{A character vector of column names. Please ensure that you only choose column names of non-numeric columns or numeric columns with few values.
Combinations that generate too few (<30)}

\item{response}{The column containing the response variable.}

\item{functions}{A (named) list of functions to be used to generate statistics. Will take a vector and should return a scalar, e.g. mean / sd.
If names are provided, the name will be prepended to the generate column. If they are not provided, gen<index of function>_ will be prepended.}

\item{too_few_observations_cutoff}{An integer denoting the minimum required observations for a combination of values in statistics_col to be used.
If not enough observations are present, the statistics will be generated on the entire response column. Default: 30.}
}
\value{
A list containing the generated statistics tables and defaults per columns
}
\description{
Calculates stats based on custom functions on the response variable for each group provided in stat_cols.
}
\details{
This function will also generate default values for all generated columns that use the entire response column.
This allows us to ensure no NA values will be present in generated columns
}
