% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{feature_categorical_filter_predict}
\alias{feature_categorical_filter_predict}
\title{Filters categorical variables for new datasets}
\usage{
feature_categorical_filter_predict(data, categorical_columns, mappings,
  insufficient_occurance_marker)
}
\arguments{
\item{data}{A new dataset}

\item{categorical_columns}{The columns from \code{feature_categorical_filter}}

\item{mappings}{The mapping from \code{feature_categorical_filter}}

\item{insufficient_occurance_marker}{The marker from \code{feature_categorical_filter}}
}
\value{
The transformed dataset
}
\description{
Filters categorical variables for new datasets
}
