% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{feature_create_all_generic_stats}
\alias{feature_create_all_generic_stats}
\title{Generic function for creating statistics on the response column, based on custom columns.}
\usage{
feature_create_all_generic_stats(train,
  stat_cols = colnames(train)[purrr::map_lgl(train, is.character)], response,
  functions, interaction_level = 1, too_few_observations_cutoff = 30)
}
\arguments{
\item{train}{The train dataset, as a data.frame. Will generate the statistics only based on this dataset.}

\item{stat_cols}{A character vector of column names. Please ensure that you only choose column names of non-numeric columns}

\item{response}{The column containing the response variable.}

\item{functions}{A (named) list of functions to be used to generate statistics. Will take a vector and should return a scalar, e.g. mean / sd.
If names are provided, the name will be prepended to the generate column. If they are not provided, gen<index of function>_ will be prepended.}

\item{interaction_level}{Either a 1 or 2. Should we gather statistics only for one column, or also for combinations of columns?}

\item{too_few_observations_cutoff}{An integer denoting the minimum required observations for a combination of values in statistics_col to be used.
If not enough observations are present, the statistics will be generated on the entire response column. Default: 30.}
}
\value{
A list containing the transformed train dataset and a trained pipe.

#' @details This function will also generate default values for all generated columns that use the entire response column.
This allows us to ensure no NA values will be present in generated columns when, for instance, a new category is detected or when values are cut-off by
\code{too_few_observations_cutoff}.
}
\description{
Generic function for creating statistics on the response column, based on custom columns.
}
