% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{feature_create_predict}
\alias{feature_create_predict}
\title{Uses previous statistics results to generate columns for a new dataset.}
\usage{
feature_create_predict(data, stat_cols, tables, interaction_level, defaults)
}
\arguments{
\item{data}{A new dataset to join with the tables generated by \code{\link{feature_create_all_generic_stats}}}

\item{stat_cols}{A character vector of column names. Please ensure that you only choose column names of non-numeric columns}

\item{tables}{The tables from \code{\link{feature_create_all_generic_stats}}.}

\item{interaction_level}{Interaction level that was used to generate the tables.}

\item{defaults}{Default values for each function, with names matching each function}
}
\value{
The \code{data} dataset, with the new columns.
}
\description{
Uses previous statistics results to generate columns for a new dataset.
}
