% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_finder.R
\name{feature_finder}
\alias{feature_finder}
\title{Add features one-by-one to find good small sets of features}
\usage{
feature_finder(train, test, response, model, difference, verbose = F)
}
\arguments{
\item{train}{The training dataset}

\item{test}{The testing dataset}

\item{response}{The response column as a string}

\item{model}{A list containing at least a training function \code{.train} and a \code{.predict} function, plus optional named
parameters to the train function.

The \code{.train} function has to take a \code{data} argument that stores the training data and a \code{...} argument for the parameters.
The \code{.predict} function needs to take two arguments, where the first is the model and the second the new dataset.

You can use \code{\link{model_trainer}} as a wrapper for this list. It will also test your inputs to a certain degree.}

\item{difference}{A function that calculates the difference between labels and predictions, for example
\code{function(x,y) abs(x-y)}}

\item{verbose}{Flag indicating if intermediate updates should be printed}
}
\value{
A dataframe, containing the mean difference for train and test set, as well as the feature that was added at each step.
}
\description{
Add features one-by-one to find good small sets of features
}
\details{
This function will select features one-by-one depending on the correlation of a feature with the residuals of the train predictions to slowly build up a bigger model.
}
