% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{feature_interactions}
\alias{feature_interactions}
\title{Generates permutation interaction effects between sets of numeric variables}
\usage{
feature_interactions(train, response, columns = 10L, max_interactions = 2)
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{response}{The column containing the response variable.}

\item{columns}{Columns to use for interaction effects. Can be a character vector referencing numeric columns, or an integer larger than 2 denoting
the minimum number of unique values in any column except \code{response} for that column to be considered for interaction effect. This is intended to exclude
ordinal columns using numeric representations.}

\item{max_interactions}{The maximum number of columns that will be considered for interaction effects per variable. E.g. for the value 3, this function
will generate ALL 2-way interactions between variables and ALL 3-way interactions between variables. Caution is advised to not set this value too high. Defaults to 2.}
}
\value{
A named list containing the modified train and the used columns and column means for reproducability.
}
\description{
Generates permutation interaction effects between sets of numeric variables
}
\details{
Thanks Eduardo.
}
