% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{feature_NA_indicators}
\alias{feature_NA_indicators}
\title{Indicate which fields are NA}
\usage{
feature_NA_indicators(train, conditions = list(is.na),
  columns = colnames(train), force_column = F)
}
\arguments{
\item{train}{The train dataset, as a data.frame.}

\item{conditions}{Vector of functions / values. All functions / values will be tested against each column. If at least one function returns true or one value matches, the records is considered a missing value.}

\item{columns}{The columns to check for missing values. Can be provided as logicals, integers, or characters}

\item{force_column}{If true, always add new columns, even if no missing values were found.}
}
\value{
A list containing the transformed train dataset, a .predict function to repeat the process on new data and all parameters needed to replicate the process.
}
\description{
Takes the dataset, scans the given columns for values that have been listed as indicating NA, and adds a column indicating if this is a NA
}
