% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{feature_NA_indicators_predict}
\alias{feature_NA_indicators_predict}
\title{Indicate which fields are NA}
\usage{
feature_NA_indicators_predict(data, conditions, columns)
}
\arguments{
\item{data}{The new dataset, as a data.frame.}

\item{conditions}{Vector of functions / values. All functions / values will be tested against each column. If at least one function returns true or one value matches, the records is considered a missing value.}

\item{columns}{The columns to check for missing values. Can be provided as logicals, integers, or characters}
}
\value{
The dataset with NA's properly processed.
}
\description{
Takes the dataset, scans the given columns for values that have been listed as indicating NA, and adds a column indicating if this is a NA
}
