% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{feature_scaler}
\alias{feature_scaler}
\title{Rescales data to standardised ranges}
\usage{
feature_scaler(train, exclude_columns = character(length = 0),
  type = "[0-1]")
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{exclude_columns}{Names of columns that should be excluded from rescaling}

\item{type}{Type of rescales to perform:
\itemize{
\item \code{"[0-1]"}: rescales the columns to the [0-1] range
\item \code{"N(0,1)"}: rescales the columns to mean 0 and sd 1
}}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Rescales data to standardised ranges
}
