% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{feature_transformer_post_predict}
\alias{feature_transformer_post_predict}
\title{Uses the results of \code{\link{pipe_feature_transformer}} on new datasets}
\usage{
feature_transformer_post_predict(data, retransform_columns,
  lower_thresholds, upper_thresholds, retransform_functions)
}
\arguments{
\item{data}{A new dataset}

\item{retransform_columns}{The same columns used to obtain these selected retransform_functions}

\item{lower_thresholds}{A vector of lower treshold used by \code{\link{uniroot}}, used to approximate the inverse function.}

\item{upper_thresholds}{A vector of upper treshold used by \code{\link{uniroot}}, used to approximate the inverse function.}

\item{retransform_functions}{Result of \code{\link{pipe_feature_transformer}}}
}
\value{
Returns the dataset with retransformed columns
}
\description{
Uses the results of \code{\link{pipe_feature_transformer}} on new datasets
}
