% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_finder.R
\name{find_best_models}
\alias{find_best_models}
\title{Select top models from the find_model function}
\usage{
find_best_models(train, find_model_result, metric, higher_is_better,
  per_model = F, top_n = 1, aggregate_func = NA)
}
\arguments{
\item{train}{The training dataset}

\item{find_model_result}{Result from the find_model function}

\item{metric}{Target metric, as a string}

\item{higher_is_better}{Logical indicating if the results should be sorted from high to low.}

\item{per_model}{Logical indicating if we should take N models per model type or in total}

\item{top_n}{The top n models to return}

\item{aggregate_func}{Aggregation function to apply. Useful if you choose more than 1 model. Set this to NA to skip it.}
}
\value{
A pipeline to create predictions for new data
}
\description{
Select top models from the find_model function
}
