% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_finder.R
\name{find_expand_results}
\alias{find_expand_results}
\title{Expands the params column of the result of \code{find_model}}
\usage{
find_expand_results(find_model_result)
}
\arguments{
\item{find_model_result}{The result from \code{find_model}}
}
\value{
A dataframe with the params column expanded
}
\description{
Expands the params column of the result of \code{find_model}
}
