% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_finder.R
\name{find_template_formula_and_data}
\alias{find_template_formula_and_data}
\title{A convenient wrapper function for find_model for models that have a formula and data argument, like lm}
\usage{
find_template_formula_and_data(response, training_function, ...)
}
\arguments{
\item{response}{The response column, as a string.}

\item{training_function}{The function that trains the algorithm, such as \code{\link{lm}}}

\item{...}{Named vectors of parameters to explore.}
}
\value{
A list that's compatible with the \code{\link{find_model}} function
}
\description{
A convenient wrapper function for find_model for models that have a formula and data argument, like lm
}
\examples{
# uses svm function from library e1071
find_template_formula_and_data(response = "response", training_function = svm,
    type = c("eps-regression", "nu-regression"),
    kernel = c("radial"), gamma = c(.001, .002, .003, .0008))
}
