% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_finder.R
\name{find_xgb}
\alias{find_xgb}
\title{A convenient wrapper function for find_model for xgboost}
\usage{
find_xgb(response, ...)
}
\arguments{
\item{response}{The response column, as a string.}

\item{...}{Named vectors of parameters to explore.}
}
\value{
A list that's compatible with the \code{\link{find_model}} function
}
\description{
A convenient wrapper function for find_model for xgboost
}
\examples{
find_xgb(response = "response", nrounds = c(30, 50), eta = c(.015),
         max_depth = c(2), colsample_bytree = c(.7), lambda = c(.1),
         subsample = c(.7), base_score = c(100))
}
