% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_removal.R
\name{greedy_max_independent_set}
\alias{greedy_max_independent_set}
\title{Finds a maximum independent set using greedy search}
\usage{
greedy_max_independent_set(graph)
}
\arguments{
\item{graph}{igraph to look for this set}
}
\value{
A set of maximum independent points
}
\description{
Finds a maximum independent set using greedy search
}
