% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_all}
\alias{impute_all}
\title{Impute multiple missing columns using lm, mean, or xgboost, and perform imputation}
\usage{
impute_all(train, columns, na_function = is.na, exclude_columns,
  type = "lm", controls = NA, verbose = F)
}
\arguments{
\item{train}{Train dataset.}

\item{columns}{The columns to impute, as strings.}

\item{na_function}{A function which returns TRUE when a value is missing and FALSE otherwise. Will apply this function to each column. Must take one column vector as input.}

\item{exclude_columns}{Columns that should not be used in imputation. If lm is chosen, this will always include \code{columns}. Should be strings.}

\item{type}{lm, mean, or xgboost.}

\item{controls}{Controls for xgboost, if needed. Default to NA.}

\item{verbose}{Whether xgboost should print anything.}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Impute multiple missing columns using lm, mean, or xgboost, and perform imputation
}
