% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_all_trees}
\alias{impute_all_trees}
\title{Impute multiple missing columns using mean / mode, lm or xgboost}
\usage{
impute_all_trees(train, columns = colnames(train)[purrr::map_lgl(train,
  na_function)], na_function = is.na, exclude_columns, type = "lm",
  controls = NA, verbose = T)
}
\arguments{
\item{train}{train dataset.}

\item{columns}{the columns to impute.}

\item{na_function}{what value represents missing values. Defaults to is.na. Either:
\itemize{
\item A function, which returns TRUE when a value is missing and FALSE otherwise. Will apply this function to each column. Must take one column vector as input.
\item A list of functions, all in the same format as above. Must be as long as \code{columns}.
}}

\item{exclude_columns}{what columns should not be used as predictors? Can be integers, logicals or column names.}

\item{type}{mean, tree, or xgboost.}

\item{controls}{controls for the decision tree or xgboost, if needed. Default to NA.}

\item{verbose}{whether xgboost should print anything}
}
\value{
A list of models to impute NA's in the columns.
}
\description{
Impute multiple missing columns using mean / mode, lm or xgboost
}
