% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_predict_all}
\alias{impute_predict_all}
\title{Use the models from \code{impute_all} to impute the selected columns in data.}
\usage{
impute_predict_all(data, columns, na_function, models, exclude_columns,
  verbose = F)
}
\arguments{
\item{data}{data dataset to impute.}

\item{columns}{the columns to impute.}

\item{na_function}{Function which returns TRUE when a value is missing and FALSE otherwise.}

\item{models}{The models generated by \code{impute_all_models}}

\item{exclude_columns}{what columns should not be used as predictors? Can be integers, logicals or column names.}

\item{verbose}{whether xgboost should print anything}
}
\value{
The same dataset as the imputed, but with NA values in the selected columns replaced by their estimated values.
}
\description{
Use the models from \code{impute_all} to impute the selected columns in data.
}
