% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_tree}
\alias{impute_tree}
\title{Generate a model to impute missing data in a column}
\usage{
impute_tree(data, column, NA_value = is.na, exclude_columns,
  train_fraction = 0.9, controls = NA, type = "xgboost")
}
\arguments{
\item{data}{The dataset, as a data.frame.}

\item{column}{The column to be imputed. Should be a string.}

\item{NA_value}{A function to define what a NA is for this column which returns TRUE when a value is missing and FALSE otherwise. Will take one column vector as input.}

\item{exclude_columns}{Columns that shouldn't be included in the imputation.}

\item{train_fraction}{The fraction of training data that should be used for imputation vs verification. The fraction will be taken from the data that doesn't have missing values. The remaining data will be used as a test set}

\item{controls}{Either:
\itemize{
\item NA for defaults
\item A list for params from xgboost. Should always contain at least \code{nrounds}.
}}

\item{type}{The type of algorithm to use for imputation. Options: mean, lm, and xgboost. Mean will calculate the mean for numeric sets and the mode for non-numerics.}
}
\value{
The model built based on \code{type} If train_fraction is below 1, will also return the prediction results on the train and test set.
}
\description{
We don't recommend to impute categorical variables with many levels.
}
