% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{invoke.pipe}
\alias{invoke.pipe}
\title{Applies a pipe to new data}
\usage{
\method{invoke}{pipe}(pipe, data)
}
\arguments{
\item{pipe}{The pipe to be applied}

\item{data}{A new dataframe}
}
\value{
The transformed dataset
}
\description{
Applies a pipe to new data
}
