% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{invoke}
\alias{invoke}
\title{Generic function to apply either a pipe or pipeline to new data}
\usage{
invoke(x, data, ...)
}
\arguments{
\item{x}{The pipe / pipeline to be applied}

\item{data}{A new dataframe}

\item{...}{Arguments for other \code{invoke} functions}
}
\value{
The transformed dataset
}
\description{
Generic function to apply either a pipe or pipeline to new data
}
