% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{is.pipe}
\alias{is.pipe}
\title{Tests if an object inherits from pipe}
\usage{
is.pipe(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
A boolean indicating if \code{x} inherits from \code{\link{pipe}}
}
\description{
Tests if an object inherits from pipe
}
