% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{is.pipeline}
\alias{is.pipeline}
\title{Tests if an object inherits from pipeline}
\usage{
is.pipeline(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
A boolean indicating if \code{x} inherits from \code{\link{pipeline}}
}
\description{
Tests if an object inherits from pipeline
}
