% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_finder.R
\name{model_trainer}
\alias{model_trainer}
\title{Wrapper function for model inputs to \code{find_model}}
\usage{
model_trainer(.train, .predict, ...)
}
\arguments{
\item{.train}{The training function. This function has to take a \code{data} argument that stores the training data and a \code{...} argument for the parameters.}

\item{.predict}{The prediction function. This function needs to take two arguments, where the first is the model and the second the new dataset.}

\item{...}{Other parameters to be passed to the training function.}
}
\value{
A list for use as a model input with \code{\link{find_model}}
}
\description{
Wrapper function for model inputs to \code{find_model}
}
\examples{
d <- data.frame(x = 1:10, y = 1:10 + rnorm(10, sd = .3))
model_trainer(
    .train = function(data, ...) lm(data = data, formula = y ~ ., ...),
    .predict = function(x,y) predict.lm(x,y),
    x = FALSE
)
}
