% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{NA_indicators_predict}
\alias{NA_indicators_predict}
\title{Indicate which fields are NA}
\usage{
NA_indicators_predict(data, condition, columns)
}
\arguments{
\item{data}{The new dataset, as a data.frame.}

\item{condition}{Function to test if a value is missing. Should return true when a value is missing and false otherwise.}

\item{columns}{The columns to check for missing values. Can be provided as logicals, integers, or characters}
}
\value{
The dataset with NA's properly processed.
}
\description{
Takes the dataset, scans the given columns for values that have been listed as indicating NA, and adds a column indicating if this is a NA
}
