% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{pipe_categorical_filter}
\alias{pipe_categorical_filter}
\title{Remove values from categorical variables that do not occur often enough}
\usage{
pipe_categorical_filter(train, response,
  insufficient_occurance_marker = "insignificant_category",
  categorical_columns = colnames(train)[purrr::map_lgl(train,
  is.character)], threshold_function = function(data) 30)
}
\arguments{
\item{train}{The train dataset, as a data.frame or data.table.}

\item{response}{The response column, as a string. Will only be used to ensure this is not included in the \code{categorical_columns} variable.}

\item{insufficient_occurance_marker}{The value to substitute when another value in the categorical column doesn't occur often enough.
This can be a string, integer or numeric.}

\item{categorical_columns}{The columns to apply the filter over. Should be a character vector.}

\item{threshold_function}{A function that takes the train dataset and produces a scalar value: the minimum number of times a value has to occur if it should be preserved.}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Remove values from categorical variables that do not occur often enough
}
\details{
Be careful: if for instance only one value gets substituted in a column, then the \code{insufficient_occurance_marker} value will just replace that one, preserving the problem.
}
