% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_check}
\alias{pipe_check}
\title{Create a pipeline step that learns what the data looks like}
\usage{
pipe_check(train, response, on_missing_column = c("error", "add")[1],
  on_extra_column = c("remove", "error")[1],
  on_type_error = c("ignore", "error")[1])
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{response}{The response variable. Will be used as an optional column name.
Does not have to exist in the train dataset (useful when it is added in later in the pipeline)}

\item{on_missing_column}{What to do when a new dataset misses columns.
Either "error", which causes an error, or "add", which adds the missing columns with only NA's filled in.}

\item{on_extra_column}{What to do when a new dataset has extra column.s
Either "error", which causes an error, or "remove", which removes the extra columns.}

\item{on_type_error}{What to do when a new dataset causes warnings or errors on casting columns to new types.
Either "error", which causes an error, or "ignore", which ignores warnings but will still allow errors to propagate.}
}
\value{
A list of the train dataset and a pipe to be used on new data.
}
\description{
Create a pipeline step that learns what the data looks like
}
