% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{pipe_clustering}
\alias{pipe_clustering}
\title{Add cluster labels to a training set}
\usage{
pipe_clustering(train, cluster_column = "cluster",
  exclude_columns = character(0), k = 4, metric = "euclidean")
}
\arguments{
\item{train}{The train dataset, as a data.frame.}

\item{cluster_column}{Name for the new cluster label column}

\item{exclude_columns}{Columns to be excluded from the clustering. Should be provided as a character vector.
Non-numerical columns will automatically be excluded}

\item{k}{The number of clusters.}

\item{metric}{The distance metric used. Currently only 'euclidean' and 'manhattan' are supported.}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Add cluster labels to a training set
}
