% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe_dplyr}
\alias{pipe_dplyr}
\title{Wrapper function to turn a dplyr function into a pipeline element}
\usage{
pipe_dplyr(dplyr_function, stop_on_missing_names = F)
}
\arguments{
\item{dplyr_function}{A function like select_ or mutate_ from dplyr}

\item{stop_on_missing_names}{Flag indicating if the pipeline should stop if ... arguments aren't named. Useful for mutate, not so much for select or group_by}
}
\value{
A wrapper function for the specified dplyr_function
}
\description{
Wrapper function to turn a dplyr function into a pipeline element
}
