% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{pipe_feature_interactions}
\alias{pipe_feature_interactions}
\title{Generates permutation interaction effects between sets of numeric variables}
\usage{
pipe_feature_interactions(train, response, columns = 10L,
  max_interactions = 2)
}
\arguments{
\item{train}{The train dataset, as a data.frame or data.table. Data.tables may be changed by reference.}

\item{response}{The column containing the response variable.}

\item{columns}{Columns to use for interaction effects. Can be a character vector referencing numeric columns, or an integer larger than 2 denoting
the minimum number of unique values in any column except \code{response} for that column to be considered for interaction effects. This is intended to exclude
ordinal columns using numeric representations.}

\item{max_interactions}{The maximum number of columns that will be considered for interaction effects per variable. E.g. for the value 3, this function
will generate ALL 2-way interactions between variables and ALL 3-way interactions between variables. Caution is advised to not set this value too high. Defaults to 2.}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Generates permutation interaction effects between sets of numeric variables
}
\details{
Thanks Eduardo.
}
