% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{pipe_feature_transformer}
\alias{pipe_feature_transformer}
\title{Applies different transformations to each numeric feature and selects the one with the highest correlation to the response per feature}
\usage{
pipe_feature_transformer(train, response, transform_columns,
  missing_func = is.na, transform_functions = list(sqrt, log,
  function(x) x^2), retransform_columns)
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{response}{String denoting the name of the column that should be used as the response variable. Mandatory}

\item{transform_columns}{Columns to consider for transformation.}

\item{missing_func}{A function that determines if an observation is missing, e.g. \code{is.na}, \code{is.nan}, or \code{function(x) x == 0}.
Defaults to \code{is.na}. Set to NA to skip NA feature generation and imputation.}

\item{transform_functions}{A list of function used to transform features. Should be one-to-one transformations like sqrt or log.}

\item{retransform_columns}{Columns that should be retransformed later on. If this is set to one or more column names, this function will generate a
numerical approximation of the inverse of the optimal tranformation function. This pipe will be returned as a separate list entry.}
}
\value{
A list containing the transformed train dataset and a trained pipe. If \code{retransform_columns} was set, the reverting pipe will also be provided.
}
\description{
Applies different transformations to each numeric feature and selects the one with the highest correlation to the response per feature
}
