% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{pipe_NA_indicators}
\alias{pipe_NA_indicators}
\title{Indicate which fields are NA}
\usage{
pipe_NA_indicators(train, condition = is.na, columns = colnames(train),
  force_column = F)
}
\arguments{
\item{train}{The train dataset, as a data.frame.}

\item{condition}{Function to test if a value is missing. Should return true when a value is missing and false otherwise.}

\item{columns}{Names of the columns to check for missing values.}

\item{force_column}{If true, always add new columns, even if no missing values were found.}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Takes the dataset, scans the given columns for values that have been listed as indicating NA, and adds a column indicating if this is a NA
}
\details{
Generated columns will be of type logical.
}
