% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{pipe_one_hot_encode}
\alias{pipe_one_hot_encode}
\title{Train one-hot encoding}
\usage{
pipe_one_hot_encode(train, columns = colnames(train)[purrr::map_lgl(train,
  function(x) return(!(is.numeric(x) || is.logical(x))))], stat_functions,
  response, use_pca = FALSE, pca_tol = 0.1)
}
\arguments{
\item{train}{Dataframe to determine one-hot-encoding.}

\item{columns}{Columns from \code{train} to use for one-hot-encoding. Will automatically check if theses are column names in \code{train}}

\item{stat_functions}{A (named) list of functions for when you want to use mean-encoding. Don't set it if you want to do regular one-hot encoding.
Any function that return a single value from a scalar would do (e.g. quantile, sd).}

\item{response}{String denoting the name of the column that should be used as the response variable. Mandatory}

\item{use_pca}{Whether PCA transformation is required.}

\item{pca_tol}{The \code{tol} of \code{\link[stats]{prcomp}}}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Train one-hot encoding
}
