% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_removal.R
\name{pipe_remove_high_correlation_features}
\alias{pipe_remove_high_correlation_features}
\title{Removes highly correlated features whilst keeping as many as possible, using heuristics.}
\usage{
pipe_remove_high_correlation_features(train,
  exclude_columns = character(0), threshold = 0.8)
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{exclude_columns}{Columns that should not be considered for removal.}

\item{threshold}{If the correlation is between two columns is larger than this, both will be considered for removal}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Removes highly correlated features whilst keeping as many as possible, using heuristics.
}
