% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline_mutate}
\alias{pipeline_mutate}
\title{Applies mutate in a pipeline}
\usage{
pipeline_mutate(train, ...)
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{...}{Variables to be made, as used in \code{\link[dplyr]{mutate_}}. This means you should provide arguments as strings: var = "a + b".}
}
\value{
A list of the transformed train dataset and a .predict function to be used on new data.
}
\description{
Applies mutate in a pipeline
}
