% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline_select}
\alias{pipeline_select}
\title{Applies select in a pipeline}
\usage{
pipeline_select(train, ...)
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{...}{Variables to be dropped / selected, as used in \code{\link[dplyr]{select_}}. This means you should provide arguments as strings: "col_1".}
}
\value{
A list of the transformed train dataset and a .predict function to be used on new data.
}
\description{
Applies select in a pipeline
}
