% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{preserve_columns_predict}
\alias{preserve_columns_predict}
\title{Only keep previously selected columns}
\usage{
preserve_columns_predict(data, preserved_columns)
}
\arguments{
\item{data}{New data frame to remove same columns from.}

\item{preserved_columns}{A vector of columns that should be preserved}
}
\value{
The same data frame, without single-value columns
}
\description{
Only keep previously selected columns
}
\details{
To be used by \code{\link{remove_single_value_columns}} and \code{\link{cor_remove_high_correlation_features}}
}
