% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_classifier.R
\name{range_predict}
\alias{range_predict}
\title{Use generated models and scales to create features for a new dataset}
\usage{
range_predict(data, models, column_names, scales, included_columns)
}
\arguments{
\item{data}{The new dataset, as a data.frame or data.table.}

\item{models}{List of models generated by \code{\link{pipe_range_classifier}}}

\item{column_names}{List of full column names for the newly generated features.}

\item{scales}{List of scales generated by \code{\link{pipe_range_classifier}}}

\item{included_columns}{Columns that should be used in the models.}
}
\value{
\code{data}, but including the new features as columns
}
\description{
Use generated models and scales to create features for a new dataset
}
