% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{remove_single_value_columns}
\alias{remove_single_value_columns}
\title{Remove all columns that have only a single value}
\usage{
remove_single_value_columns(train, na_function = function(x) {     F })
}
\arguments{
\item{train}{train data frame to select columns from to remove.}

\item{na_function}{A function that returns true when a value is considered missing, and should be removed. Defaults to removing no values.}
}
\value{
A list containing the transformed train dataset, a .predict function to repeat the process on new data and all parameters needed to replicate the process.
}
\description{
Remove all columns that have only a single value
}
