% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_checks.R
\name{sanity_checking}
\alias{sanity_checking}
\title{Checks for a couple of standard things to ensure the data is okay.}
\usage{
sanity_checking(data, missing_values = list(NA, "NA", "<NA>", "none", "",
  NULL, "NULL", "NaN", NaN, 0, -1),
  min_character_freq = ceiling(log(nrow(data), base = 2)),
  other_type_regexs = character(0), n_sigma = 3)
}
\arguments{
\item{data}{The dataset to test}

\item{missing_values}{A list of missing values to test for. Strings will only be tested against character columns, NA, NaN, NULL and numerical values against all columns.}

\item{min_character_freq}{How often an entry in a character column needs to occur to not be marked as an "outlier".}

\item{other_type_regexs}{Additional regexes to check for possible encoding of other variables}

\item{n_sigma}{The n in \code{[median - n*sd, median + n*sd]}, used to determine upper and lower bounds on acceptable values for numerical columns}
}
\value{
A list with 3 dataframes:
\itemize{
\item numerical_summary
\item categorical_summary
\item missing_value_table
}
}
\description{
Checks for a couple of standard things to ensure the data is okay.
}
\details{
Test the following items:
\itemize{
\item Column types are correct, ergo detect if:
  \itemize{
  \item Character columns don't contain to many numbers. (Bad interpretation of a column)
  \item Numeric columns actually represent numerical scales, e.g. no categorical encoding.
  }
\item (Simple) outlier detection per column. This will not take into account the above mentioned column checks:
  \itemize{
  \item Character columns: occurances of entries < \code{min_character_freq} and which these are.
  \item Numeric columns: occurances of values outside the range \code{[median - n*sd, median + n*sd]}. The bounds will be set within the range of each column.
  Will give this range and a flag indicating if there are any records outside of this range
  }
\item Detect possible encodings of missing values:
  \itemize{
  \item All columns: 0, -1, NA, NULL, NaN
  \item Character columns: "NA", "none", "<NA>", "NULL", "NaN", ""
  }
  Non-character/numeric values such as NA, NaN and NULL will get \code{nsv:} prepended to them (non-standard value)
\item Detect possible encodings for other date types in character columns, such as:
  \itemize{
  \item Dates
  \item Timestamps
  \item Emails
  }
}
}
