% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{segment}
\alias{segment}
\title{A simple wrapper for creating a pipe segment}
\usage{
segment(.segment, ...)
}
\arguments{
\item{.segment}{The function to train a part of the pipeline. Will be checked for being a function and taking a \code{train} argument.}

\item{...}{Other arguments to \code{.segment}.}
}
\value{
A list with .segment and the argument in ..., ready for \code{\link{pipeline}}
}
\description{
A simple wrapper for creating a pipe segment
}
\examples{
p <- segment(pipeline_mutate, a = "1")
p2 <- segment(pipeline_mutate, "-a")
}
