% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standard_column_names}
\alias{standard_column_names}
\title{Standardises column names into an oft-acceptable format}
\usage{
standard_column_names(data)
}
\arguments{
\item{data}{The dataset to standardise}
}
\value{
The dataset with standardised column names
}
\description{
Standardises column names into an oft-acceptable format
}
\examples{
data <- dplyr::data_frame(var = 0, Var = 0, camelCase = 0,
                          good_name = 0, `0none.` = 0, `bad  ` = 0,
                          `j&d` = 0, `spac ed` = 0)
standard_column_names(data)
}
