% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{build_model_package}
\alias{build_model_package}
\title{Builds a package from your pipeline}
\usage{
build_model_package(trained_pipeline, package_name = "deploymodel",
  libraries = names(utils::sessionInfo()$otherPkgs),
  tar_file = "deploy.tar.gz", extra_variables = character(0),
  may_overwrite_tar_file = F, verbose = F)
}
\arguments{
\item{trained_pipeline}{The model function from \code{find_best_models}}

\item{package_name}{The name of the package}

\item{libraries}{A list of library names required by the package. Defaults to all loaded non-base packages}

\item{tar_file}{The name and path of the final tarred package}

\item{extra_variables}{A character vector of variables names to be included in the package. Used to include extra fuctions and variables in the packages
that are not contained in any of the packages listed in \code{libraries}.}

\item{may_overwrite_tar_file}{Flag indicating if, when \code{tar_file} exists, this function is allowed to override it.}

\item{verbose}{Flag indicating if devtools functions should print anything.}
}
\value{
A logical: TRUE for success and FALSE for failure
}
\description{
Builds a package from your pipeline
}
\details{
A tip for finding out which libraries you need: start with just datapiper, build the package and check the output of building the package.
It will display which functions you use that aren't included yet.
}
