% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{build_model_package}
\alias{build_model_package}
\title{Builds a package from your pipeline}
\usage{
build_model_package(trained_pipeline, package_name = "deploymodel",
  libraries = names(utils::sessionInfo()$otherPkgs),
  tar_file = "deploy.tar.gz", extra_functions = character(0),
  prediction_precision = 8, may_overwrite_tar_file = F)
}
\arguments{
\item{trained_pipeline}{The model function from \code{find_best_models}}

\item{package_name}{The name of the package}

\item{libraries}{A list of library names required by the package. Defaults to all loaded non-base packages}

\item{tar_file}{The name and path of the final tarred package}

\item{extra_functions}{A character vector of function names to be included in the package.}

\item{prediction_precision}{Number of digits to be returned by the prediction}

\item{may_overwrite_tar_file}{Flag indicating if, when \code{tar_file} exists, this function is allowed to override it.}
}
\value{
A logical: TRUE for success and FALSE for failure
}
\description{
Builds a package from your pipeline
}
\details{
A tip for finding out which libraries you need: start with just datapiper, build the package and check the output of building the package.
It will display which functions you use that aren't included yet.
}
