% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{feature_one_hot_encode_predict}
\alias{feature_one_hot_encode_predict}
\title{Apply one-hot encoding}
\usage{
feature_one_hot_encode_predict(data, one_hot_parameters, use_pca, pca,
  stat_transformer, generated_columns)
}
\arguments{
\item{data}{A new dataframe}

\item{one_hot_parameters}{The \code{one_hot_parameters} element from the result of \code{\link{pipe_one_hot_encode}}}

\item{use_pca}{Whether PCA transformation is required.}

\item{pca}{The PCA result from \code{\link{pipe_one_hot_encode}}. Required iff \code{use_pca} is \code{TRUE}}

\item{stat_transformer}{Stat transformer function from \code{\link{pipe_one_hot_encode}}, or NA (for normal one-hot encoding)}

\item{generated_columns}{Names of generated columns}
}
\value{
The test dataframe with the required columns one-hot encoded. Only required if stat_transformer is used.
}
\description{
Apply one-hot encoding
}
