% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{feature_scaler}
\alias{feature_scaler}
\title{Rescales data to standardised ranges}
\usage{
feature_scaler(train, response,
  columns = colnames(train)[purrr::map_lgl(train, is.numeric)],
  type = "[0-1]")
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{response}{String denoting the name of the column that should be used as the response variable. Mandatory}

\item{columns}{The columns that need to be rescaled.}

\item{type}{Type of rescales to perform:
\itemize{
\item \code{"[0-1]"}: rescales the columns to the [0-1] range
\item \code{"N(0,1)"}: rescales the columns to mean 0 and sd 1
}}
}
\value{
A list containing the transformed train dataset, a .predict function to repeat the process on new data and all parameters needed to replicate the process.
}
\description{
Rescales data to standardised ranges
}
