% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{feature_transformer_predict}
\alias{feature_transformer_predict}
\title{Uses the results of \code{\link{feature_transformer}} on new datasets}
\usage{
feature_transformer_predict(data, transform_columns, transform_functions)
}
\arguments{
\item{data}{A new dataset}

\item{transform_columns}{The same columns used to obtain these selected transform_functions}

\item{transform_functions}{Result of \code{\link{feature_transformer}}}
}
\value{
Returns
}
\description{
Uses the results of \code{\link{feature_transformer}} on new datasets
}
