% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipe}
\alias{pipe}
\title{Creates a pipe object out of a function and a list of arguments}
\usage{
pipe(.function, ...)
}
\arguments{
\item{.function}{A function to repeat transformations from a trained pipeline on new data. Will be checked for being a function and taking a \code{data} argument.}

\item{...}{Other arguments to \code{.function}. Non-named arguments are only accepted if \code{.function} takes a \code{...} argument. Named arguments will be checked to
be in the argument list of \code{.function}.}
}
\value{
A pipe object with two entries:
\itemize{
\item \code{predict_function}: A function to repeat transformations from a trained pipeline on new data.
\item \code{args}: Arguments for \code{predict_function}
}
}
\description{
Creates a pipe object out of a function and a list of arguments
}
