% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_features.R
\name{pipe_pca}
\alias{pipe_pca}
\title{Apply PCA to a subset of the columns in a dataset}
\usage{
pipe_pca(train, columns, pca_tol = 0.1, keep_old_columns = F)
}
\arguments{
\item{train}{Data frame containing the train data.}

\item{columns}{Columns to use in the PCA transformation.}

\item{pca_tol}{The \code{tol} of \code{\link[stats]{prcomp}}}

\item{keep_old_columns}{Flag indicating if columns used to perform the PCA transformation should be kept.}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Apply PCA to a subset of the columns in a dataset
}
