% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_generation.R
\name{pipe_remove_single_value_columns}
\alias{pipe_remove_single_value_columns}
\title{Remove all columns that have only a single value}
\usage{
pipe_remove_single_value_columns(train, na_function = function(x) {     F
  })
}
\arguments{
\item{train}{The train dataset, as a data.frame or data.table. Data.tables may be changed by reference.}

\item{na_function}{A function that returns true when a value is considered missing, and should be removed. Defaults to removing no values.}
}
\value{
A list containing the transformed train dataset and a trained pipe.
}
\description{
Remove all columns that have only a single value
}
